tweaked status page display
