+++ status.js
@@ -112,7 +112,7 @@ module.exports = function(io) {
                     });
                 },
                 last_added: function(cb) {
-                    db.matches.findOne({}, {
+                    db.matches.find({}, {
                         sort: {
                             match_seq_num: -1
                         },
@@ -121,13 +121,13 @@ module.exports = function(io) {
                             start_time: 1,
                             duration: 1
                         },
-                        limit: 1
-                    }, function(err, match) {
-                        cb(err, match);
+                        limit: 5
+                    }, function(err, matches) {
+                        cb(err, matches);
                     });
                 },
                 last_parsed: function(cb) {
-                    db.matches.findOne({
+                    db.matches.find({
                         parse_status: 2
                     }, {
                         sort: {
@@ -138,9 +138,9 @@ module.exports = function(io) {
                             start_time: 1,
                             duration: 1
                         },
-                        limit: 1
-                    }, function(err, match) {
-                        cb(err, match);
+                        limit: 5
+                    }, function(err, matches) {
+                        cb(err, matches);
                     });
                 }
             },
